Create Building class to instanciate new buildings
